requirements.txt
# Core OCR
pytesseract>=0.3.10
Pillow>=9.0.0

# Image processing
opencv-python>=4.7.0
numpy>=1.24.0
scikit-image>=0.20.0
imutils>=0.5.4

# Augmentation
albumentations>=1.3.0

# Evaluation metrics
jiwer>=3.0.0
editdistance>=0.6.2

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Annotation (Label Studio)
label-studio>=1.7.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
loguru>=0.7.0
rich>=13.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0
